[hotfix] Split too large file CheckpointCoordinatorTest.java into several small files
